requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pymongo==4.14.0
motor==3.7.1
python-dotenv==1.1.1
semanticscholar==0.10.0
aiofiles==24.1.0
python-multipart==0.0.20